% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandisBarrettGalvin2013.R
\docType{data}
\name{LandisBarrettGalvin2013}
\alias{LandisBarrettGalvin2013}
\title{LandisBarrettGalvin2013 data}
\format{
An object of class data.frame.
}
\source{
\doi{10.1037/a0033410}
}
\usage{
LandisBarrettGalvin2013
}
\description{
The data, taken from \insertCite{lbg13}{ANOFA}, is a dataset where the
participants (n = 553) are classified according to two factors,
first, how  modalities of care in a family
medicine residency program were given. The possible cases were \verb{Collocated Behavioral Health service} (CBH), a \verb{Primary-Care  Behavioral Health service} (PBH) and a \verb{Blended Model} (BM).
Second, how a patient’s care was financed:
\code{Medicare} (MC), \code{Medicaid} (MA), a \verb{mix of Medicare/Medicaid}
(MC/MA), \verb{Personal insurance} (PI), or \code{Self-paid} ($P). This
design therefore has 5 x 3 = 15 cells. It was thoroughly examined
in \insertCite{s15}{ANOFA} and analyzed in \insertCite{lc23b}{ANOFA}.
}
\examples{

# running the anofa
L <- anofa( obsfreq ~ provider * program, LandisBarrettGalvin2013)

# getting a plot
anofaPlot(L)

# the G table shows a significant interaction
summary(L)

# getting the simple effect
e <- emFrequencies(L, ~ program | provider ) 

## Getting some contrast by provider (i.e., on e)
f <- contrastFrequencies(e, list(
         "(PBH & CBH) vs. BM"=c(1,1,-2)/2, 
         "PBH vs. CBH"=c(1,-1,0))
     )


}
\references{
\insertAllCited{}
}
\keyword{datasets}
