% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOFA-methods.R
\name{summarize}
\alias{summarize}
\title{summarize}
\usage{
summarize(object, ...)
}
\arguments{
\item{object}{an object to summarize}

\item{...}{ignored}
}
\value{
a human-readable output as per articles.
}
\description{
\code{summarize()} provides a human-readable output of an ANOFAobject. it is
synonym of \code{summary()} (but as actions are verbs, I used a verb).
}
